fix(logging): initialize logging in logger file and move version log
