+++ src/so_vits_svc_fork/__init__.py
@@ -1,1 +1,5 @@
 __version__ = "1.4.1"
+
+from .logger import init_logger
+
+init_logger()
