+++ src/so_vits_svc_fork/utils.py
@@ -461,7 +461,7 @@ def clean_checkpoints(
     sort_by_time      --  True -> chronologically delete ckpts
                           False -> lexicographically delete ckpts
     """
-    LOG.warning("Cleaning old checkpoints...")
+    LOG.info("Cleaning old checkpoints...")
     path_to_models = Path(path_to_models)
 
     # Define sort key functions
@@ -488,7 +488,7 @@ def clean_checkpoints(
         to_delete_list = list(group_items)[:-n_ckpts_to_keep]
 
         for to_delete in to_delete_list:
-            LOG.warning(f"Removing {to_delete}")
+            LOG.info(f"Removing {to_delete}")
             to_delete.unlink()
